move a function into its own module
